+++ processor/src/test/resources/expected/com/example/arez/React4j_KeepAliveComputedArezReactComponent.java
@@ -2,9 +2,6 @@ package com.example.arez;
 
 import arez.annotations.Action;
 import arez.annotations.ArezComponent;
-import arez.annotations.Dependency;
-import arez.annotations.Observable;
-import arez.annotations.ObservableRef;
 import javax.annotation.Generated;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
@@ -16,58 +13,25 @@ import react4j.core.NativeAdapterComponent;
 import react4j.core.ReactConfig;
 
 @ArezComponent(
-    name = "ComponentWithDependency",
+    name = "KeepAliveComputedArezReactComponent",
     deferSchedule = true
 )
 @Generated("react4j.processor.ReactProcessor")
-abstract class React4j_ComponentWithDependency extends ComponentWithDependency {
+abstract class React4j_KeepAliveComputedArezReactComponent extends KeepAliveComputedArezReactComponent {
   static final ComponentConstructorFunction TYPE = getConstructorFunction();
 
   @Nonnull
   private static ComponentConstructorFunction getConstructorFunction() {
     final ComponentConstructorFunction componentConstructor = NativeReactComponent::new;
     if ( ReactConfig.enableComponentNames() ) {
-      Js.asPropertyMap( componentConstructor ).set( "displayName", "ComponentWithDependency" );
+      Js.asPropertyMap( componentConstructor ).set( "displayName", "KeepAliveComputedArezReactComponent" );
     }
     return componentConstructor;
   }
 
   @Override
-  @Observable(
-      name = "value",
-      expectSetter = false
-  )
-  protected String getValue() {
-    return null != props().getAny( "value" ) ? props().getAny( "value" ).asString() : null;
-  }
-
-  @Nonnull
-  @ObservableRef
-  protected abstract arez.Observable getValueObservable();
-
-  @Override
-  @Observable(
-      name = "model",
-      expectSetter = false
-  )
-  @Dependency
-  protected ComponentWithDependency.Model getModel() {
-    return null != props().getAny( "model" ) ? props().getAny( "model" ).cast() : null;
-  }
-
-  @Nonnull
-  @ObservableRef
-  protected abstract arez.Observable getModelObservable();
-
-  @Override
   @Action
   protected void reportPropsChanged(@Nullable final JsPropertyMap<Object> nextProps) {
-    if ( !Js.isTripleEqual( props().get( "value" ), null == nextProps ? null : nextProps.get( "value" ) ) ) {
-      getValueObservable().reportChanged();
-    }
-    if ( !Js.isTripleEqual( props().get( "model" ), null == nextProps ? null : nextProps.get( "model" ) ) ) {
-      getModelObservable().reportChanged();
-    }
   }
 
   @Override
@@ -90,14 +54,14 @@ abstract class React4j_ComponentWithDependency extends ComponentWithDependency {
         @Nonnull JsPropertyMap<Object> arg1);
   }
 
-  private static final class NativeReactComponent extends NativeAdapterComponent<ComponentWithDependency> implements Lifecycle {
+  private static final class NativeReactComponent extends NativeAdapterComponent<KeepAliveComputedArezReactComponent> implements Lifecycle {
     NativeReactComponent(@Nullable final JsPropertyMap<Object> props) {
       super( props );
     }
 
     @Override
-    protected ComponentWithDependency createComponent() {
-      return new Arez_React4j_ComponentWithDependency();
+    protected KeepAliveComputedArezReactComponent createComponent() {
+      return new Arez_React4j_KeepAliveComputedArezReactComponent();
     }
 
     @Override